Se agregan comentarios a la funcion auth, resolver error de login
